support for mocking client (UI test)
