+++ internal/config/config.go
@@ -6,6 +6,7 @@ type Config struct {
 	ResourceGroupName string // specified via CLI
 	Logfile           string `env:"AZTFY_LOGFILE" default:""`
 	Debug             bool   `env:"AZTFY_DEBUG" default:"false"`
+	MockClient        bool   `env:"AZTFY_MOCK_CLIENT" default:"false"`
 }
 
 func NewConfig(rg string) (*Config, error) {
